fix: cleanup and deletion of unused code
